# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Joke App API",
    "description": "Auto-generated Swagger UI for Joke App",
    "version": "1.0.0"
  },
  "host": "192.168.100.177:8010",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/jokes/": {
      "get": {
        "tags": [
          "Jokes"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Jokes"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "cat_id",
            "in": "formData",
            "required": true,
            "type": "number",
            "description": "Category Id"
          },
          {
            "name": "subcat_id",
            "in": "formData",
            "type": "number",
            "description": "Sub Category Id"
          },
          {
            "name": "content",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Joke Content"
          },
          {
            "name": "joke_image",
            "in": "formData",
            "type": "file",
            "description": "Joke Image"
          },
          {
            "name": "speaker_image",
            "in": "formData",
            "type": "file",
            "description": "Speaker Image"
          },
          {
            "name": "receiver_image",
            "in": "formData",
            "type": "file",
            "description": "Receiver Image"
          },
          {
            "name": "image_answer",
            "in": "formData",
            "type": "file",
            "description": "Answer Image"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/jokes/{id}": {
      "get": {
        "tags": [
          "Jokes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Jokes"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cat_id",
            "in": "formData",
            "required": true,
            "type": "number",
            "description": "Category Id"
          },
          {
            "name": "subcat_id",
            "in": "formData",
            "type": "number",
            "description": "Sub Category Id"
          },
          {
            "name": "content",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "Joke Content"
          },
          {
            "name": "joke_image",
            "in": "formData",
            "type": "file",
            "description": "Joke Image"
          },
          {
            "name": "speaker_image",
            "in": "formData",
            "type": "file",
            "description": "Speaker Image"
          },
          {
            "name": "receiver_image",
            "in": "formData",
            "type": "file",
            "description": "Receiver Image"
          },
          {
            "name": "image_answer",
            "in": "formData",
            "type": "file",
            "description": "Answer Image"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Jokes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}